requirements.txt
lfdocs-conf
sphinx>=1.6.5,<1.7
sphinx_bootstrap_theme>=0.4.11
robotframework
graphviz
